Rename helpers/commonmark into helpers/markdown for consistency
